feat(investment): 전략 상세 팝업 모달 + 비교 결과표 용어 설명 팝업
[전략 상세 모달]
- PresetDetailView 컴포넌트 분리 (페이지/모달 양쪽 재사용, variant prop)
- 페이지 라우트(page.tsx)는 PresetDetailView를 감싸는 얇은 래퍼로 단순화
- 비교 페이지의 ⓘ 버튼 클릭 시 새 탭 대신 PresetDetailModal 오버레이로 표시
  · ESC 키로 닫기, 배경 클릭으로 닫기, body 스크롤 잠금
  · 추천 카드의 전략명 클릭도 모달 오픈으로 변경

[비교 결과표 용어 설명 팝업]
- METRIC_GLOSSARY 사전: 순위/수익률/B&H/승률/거래/Sharpe/MDD/PF
  · 각 항목에 본문 설명 + 계산식 포함
- 결과표 헤더의 각 컬럼 라벨 옆에 ⓘ 버튼 → GlossaryModal 표시
- ESC/배경 클릭으로 닫기
